# openapi.yaml
openapi: 3.1.0
info:
  title: Songless Multiplayer API
  version: 0.1.0
  description: |
    REST + WebSocket API for the multiplayer Songless variant.

    ## Auth model
    All non-auth endpoints require a session cookie (`session=<token>`).
    The cookie is set by `POST /auth/login` and cleared by `POST /auth/logout`.

    ## Gameplay model (Heardle-style)
    Each round, every non-picker player progresses through guess brackets
    independently. At bracket index `i`, the player can play up to
    `guess_brackets_seconds[i]` seconds of audio. Each attempt the player
    either:
      - Guesses correctly → round ends for that player; they score
        `len(brackets) - i` points (default scale: 6/5/4/3/2/1).
      - Guesses incorrectly → bracket auto-advances by one.
      - Skips → bracket auto-advances by one (no guess recorded).
    A player is "exhausted" when they wrong-guess or skip past the last
    bracket — they finish the round with 0 points. The round ends when every
    non-picker has either guessed correctly or exhausted brackets, or when
    a hard server-side timeout elapses.

    ## Realtime (WebSocket: `/rooms/{code}/ws`)
    Server → client events (`{ "type": ..., "payload": ... }`):
      - `room_state`           — full snapshot, sent on connect / after any change
      - `player_joined`        — `{ user }`
      - `player_left`          — `{ user_id }`
      - `player_disconnected`  — `{ user_id }`
      - `settings_updated`     — `{ settings }`
...

...

...
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    User:
      type: object
      required: [id, username]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
    RoomStatus:
      type: string
      enum: [lobby, selecting, playing, results]
    RoomSettings:
      type: object
      properties:
        min_popularity:
          type: integer
          minimum: 0
          maximum: 100
          description: Minimum Spotify popularity (0-100).
        required_artists:
          type: array
          items: { type: string }
          description: Spotify artist IDs. If non-empty, track must be by one of these.
        songs_per_player:
          type: integer
          minimum: 1
          maximum: 10
        guess_brackets_seconds:
          type: array
          items: { type: number, minimum: 0 }
          minItems: 1
          description: |
            Strictly-ascending elapsed-time markers (in seconds) at which the score
            decreases. The last value is also the round's hard time limit.
            Default: [0.5, 1, 2.5, 5, 15, 30].
        album_art_enabled:
          type: boolean
          description: Show the song's album cover during the round.
        album_art_unblur:
          type: boolean
          description: |
            When enabled, the album cover is blurred at round start and progressively
            unblurred as each guess bracket elapses. Has no effect if
            `album_art_enabled` is false.
        round_intermission_seconds:
          type: integer
          minimum: 0
          maximum: 30
          description: |
            Pause between rounds during which the song is revealed and the
            leaderboard movement is animated. 0 disables the intermission.
        post_game_delay_seconds:
          type: integer
          minimum: 0
          description: Cooldown after results before leader can restart.
    RoomPlayer:
      type: object
      required: [user, score, connected, songs_submitted]
      properties:
        user: { $ref: "#/components/schemas/User" }
        score: { type: integer }
        connected: { type: boolean }
        songs_submitted:
          {
            type: integer,
            description: Count only; titles withheld until reveal.,
          }
    Room:
      type: object
      required: [code, leader_id, status, settings, players]
      properties:
        code: { type: string }
        leader_id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/RoomStatus" }
        settings: { $ref: "#/components/schemas/RoomSettings" }
        players:
          type: array
          items: { $ref: "#/components/schemas/RoomPlayer" }
        current_round_id:
          type: string
          format: uuid
          nullable: true
...
    GuessResult:
      type: object
      required: [correct, points]
      properties:
        correct: { type: boolean }
        points: { type: integer }
        bracket_seconds: { type: number, nullable: true }
...
  responses:
    Unauthorized:
      description: Not logged in
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Room or resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
security:
  - sessionCookie: []
paths:
  /auth/register:
    post:
      summary: Create a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, minLength: 3, maxLength: 32 }
                password: { type: string, minLength: 8 }
      responses:
        "201":
          description: Created and logged in (sets session cookie)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "409":
          description: Username taken
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/login:
    post:
      summary: Log in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Logged in (sets session cookie)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /me:
    get:
      summary: Current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /rooms:
    post:
      summary: Create a new room (caller becomes leader)
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /rooms/{code}/guess:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
    post:
      summary: Submit a guess for the caller's current bracket
      description: |
        The player picks a track from a Deezer search and submits its id. The
        backend resolves the track via Deezer and compares its normalized
        title and artist against the round's stored values. Scoring uses the
        player's server-tracked `bracket_index` at the moment the guess
        arrives. A wrong guess auto-advances the player to the next bracket;
        if they were already at the last bracket, they finish exhausted with
        0 points.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [round_id, guessed_track_id]
              properties:
                round_id: { type: string, format: uuid }
                guessed_track_id: { type: string }
      responses:
        "200":
          description: Guess recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GuessResult" }
        "409":
          description: Round not active, you've already finished, or you are the picker
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /rooms/{code}/skip:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
    post:
      summary: Skip the current bracket without guessing
      description: |
        Advances the caller to the next bracket. If already at the last bracket,
        the caller finishes the round exhausted with 0 points.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [round_id]
              properties:
                round_id: { type: string, format: uuid }
      responses:
        "200":
          description: Bracket advanced or player finished
          content:
            application/json:
              schema:
                type: object
                properties:
                  bracket_index: { type: integer }
                  finished: { type: boolean }
        "409":
          description: Round not active, you've already finished, or you are the picker
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /rooms/{code}/rounds/{round_id}/audio:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
      - {
          name: round_id,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      summary: Stream a slice of the round's audio (length determined by caller's bracket)
      description: |
        The endpoint takes no bracket parameter — it is read from the in-memory
        round state for the caller, so a client cannot ask for more audio than
        it has unlocked.

        Server-side flow:
          1. Look up the caller's `bracket_index` in the round state.
          2. Resolve cached full-length audio from Valkey
             (key: `audio:<spotify_track_id>`). On miss: fetch the preview
             from Spotify's CDN, strip ID3 tags with `mutagen`, store the
             cleaned bytes in Valkey, then continue.
          3. Slice to `guess_brackets_seconds[bracket_index]` seconds with
             pydub/ffmpeg and stream as `audio/mpeg`.

        Served with `Cache-Control: no-store` so the slice doesn't persist
        in any downstream/browser disk cache.
      responses:
        "200":
          description: Audio slice
          content:
            audio/mpeg: {}
        "403":
          description: You're the picker, or you've already finished this round
        "404": { $ref: "#/components/responses/NotFound" }
  /rooms/{code}/rounds/{round_id}/full-audio:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
      - {
          name: round_id,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      summary: Stream the full preview audio for a completed round
      description: |
        Served during the post-round intermission so all players (including
        the picker and exhausted players) can hear the song. Refuses with
        409 while the round is still in progress to avoid leaking the
        answer.
      responses:
        "200":
          description: Full preview audio
          content:
            audio/mpeg: {}
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Round has not ended yet
...